add auth for edit and delete

+++ node-rest/controllers/feed.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator/check');
 const Post = require('../models/post');
+const User = require("../models/user");
 const fs = require("fs");
 const path = require("path");
 exports.getPosts = (req, res, next) => {
@@ -13,10 +14,11 @@ exports.getPosts = (req, res, next) => {
             return Post.find().skip((currentPage - 1) * perPage).limit(perPage);
         })
         .then(posts => {
+            console.log("posts->", posts)
             res.status(200).json({
                 message: "posts fetched!!",
                 posts: posts,
-                totalItems:totalItems
+                totalItems: totalItems
             })
 
         })
@@ -59,17 +61,25 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'bob' }
+        creator: req.userId
     });
     post.save().then(result => {
         console.log(result);
+        return User.findById(req.userId);
+    }).then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    }).then(result => {
         res.status(201).json({
             message: "post created succesfully!!",
-            post: result
+            post: post,
+            creator: { _id: creator._id, name: creator.name }
         })
     }).catch(err => {
         if (!err.statusCode) {
@@ -126,6 +136,12 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("not authorized!");
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -154,7 +170,13 @@ exports.deletePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error("not authorized!");
+            error.statusCode = 403;
+            throw error;
+        }
         // check logged user
+        clearImage(post.imageUrl);
         return Post.findOneAndRemove(postId);
 
     }).then(result => {
